Fix grenade issue
If you were running forward and simply release a grenade it fired out in front of you. This fix calculates grenade velocity correctly relative to current player velocity. Only issue remaining is when player is on a moving platform like a train, it doesn't take the train velocity into account.