Use player hand for Use Action
This is now based on the dominant hand controller location, now you don't have to be looking at the button/switch etc that you want to use, the controller just has to be pointed at, and in the vicinity of, the item when you click Use. I fixed this when I noticed I couldn't raise/lower some elevators in BlueShift without looking at the switch (and therefore not being able to see what I was doing to the elevator), so that was worth fixing!